refactor: updated the primary button to include dropdown support
